src/main/java/org/algorithms/coding_patterns/educative/fast_slow_pointers/LinkedListCycle.java
package org.algorithms.coding_patterns.educative.fast_slow_pointers;

public class LinkedListCycle {
    //time - O(n), space - O(1)
    public static boolean hasCycle(ListNode head) {
        ListNode slow = head;
        ListNode fast = head;
        while (fast != null && fast.next != null) {
            fast = fast.next.next;
            slow = slow.next;
            if (slow == fast)
                return true; // found the cycle
        }
        return false;
    }
}

class ListNode {
    int value;
    ListNode next;

    ListNode(int value) {
        this.value = value;
    }
}
